Beginning to split out GUI into its own module
