+++ autopilot/core/gui.py
@@ -16,7 +16,6 @@ the method must be decorated with `@gui_event` which will call perform the updat
     This will be corrected before v1.0
 
 """
-import sys
 import typing
 import os
 import json
@@ -26,45 +25,25 @@ import time
 from collections import OrderedDict as odict
 import numpy as np
 import ast
-from PySide2 import QtGui, QtCore, QtWidgets
+from PySide2 import QtGui, QtWidgets
 import pyqtgraph as pg
 import pandas as pd
 import itertools
 import threading
-from queue import Empty, Full
 import multiprocessing as mp
-import logging
 from operator import ior
 from functools import reduce
-from collections import abc
 
 # adding autopilot parent directory to path
-import autopilot
 from autopilot.core.subject import Subject
 from autopilot import prefs
-from autopilot.stim.sound import sounds
+from autopilot.gui import _MAPS
+from autopilot.gui.gui import gui_event
 from autopilot.networking import Net_Node
-from functools import wraps
-from autopilot.utils.invoker import InvokeEvent, get_invoker
-from autopilot.core import styles
 from autopilot.core.plots import Video
 from autopilot.core.loggers import init_logger
-from autopilot.utils import plugins, registry, wiki
+from autopilot.utils import plugins, registry
 
-_MAPS = {
-    'dialog': {
-        'icon': {
-            'info': QtWidgets.QMessageBox.Information,
-            'question': QtWidgets.QMessageBox.Question,
-            'warning': QtWidgets.QMessageBox.Warning,
-            'error': QtWidgets.QMessageBox.Critical
-        },
-        'modality': {
-            'modal': QtCore.Qt.NonModal,
-            'nonmodal': QtCore.Qt.WindowModal
-        }
-    }
-}
 """
 Maps of shorthand names for objects to the objects themselves.
 
@@ -72,557 +51,9 @@ Grouped by a rough use case, intended for internal (rather than user-facing) use
 """
 
 
-def gui_event(fn):
-    """
-    Wrapper/decorator around an event that posts GUI events back to the main
-    thread that our window is running in.
-
-    Args:
-        fn (callable): a function that does something to the GUI
-    """
-    @wraps(fn)
-    def wrapper_gui_event(*args, **kwargs):
-        """
-
-        Args:
-            *args ():
-            **kwargs ():
-        """
-        QtCore.QCoreApplication.postEvent(get_invoker(), InvokeEvent(fn, *args, **kwargs))
-    return wrapper_gui_event
-
-
-class Control_Panel(QtWidgets.QWidget):
-    """A :class:`QtWidgets.QWidget` that contains the controls for all pilots.
-
-    Args:
-        subjects (dict): See :py:attr:`.Control_Panel.subjects`
-        start_fn (:py:meth:`~autopilot.core.terminal.Terminal.toggle_start`): the Terminal's
-            toggle_start function, propagated down to each :class:`~core.gui.Pilot_Button`
-        pilots: Usually the Terminal's :py:attr:`~.Terminal.pilots` dict. If not passed,
-            will try to load :py:attr:`.params.PILOT_DB`
-
-    Attributes:
-        subjects (dict): A dictionary with subject ID's as keys and
-                :class:`core.subject.Subject` objects as values. Shared with the
-                Terminal object to manage access conflicts.
-        start_fn (:py:meth:`~autopilot.core.terminal.Terminal.toggle_start`): See :py:attr:`.Control_Panel.start_fn`
-        pilots (dict): A dictionary with pilot ID's as keys and nested dictionaries
-                    containing subjects, IP, etc. as values
-        subject_lists (dict): A dict mapping subject ID to :py:class:`.subject_List`
-        layout (:py:class:`~QtWidgets.QGridLayout`): Layout grid for widget
-        panels (dict): A dict mapping pilot name to the relevant :py:class:`.Pilot_Panel`
-
-    Specifically, for each pilot, it contains
-
-    * one :class:`subject_List`: A list of the subjects that run in each pilot.
-    * one :class:`Pilot_Panel`: A set of button controls for starting/stopping behavior
-
-    This class should not be instantiated outside the context of a
-    :py:class:`~.terminal.Terminal` object, as they share the :py:attr:`.subjects` dictionary.
-
-    """
-    # Hosts two nested tab widgets to select pilot and subject,
-    # set params, run subjects, etc.
-
-    def __init__(self, subjects, start_fn, ping_fn, pilots):
-        """
-
-        """
-        super(Control_Panel, self).__init__()
-
-        self.logger = init_logger(self)
-
-        # We share a dict of subject objects with the main Terminal class to avoid access conflicts
-        self.subjects = subjects
-
-        # We get the Terminal's send_message function so we can communicate directly from here
-        self.start_fn = start_fn
-        self.ping_fn = ping_fn
-        self.pilots = pilots
-
-
-        # Make dict to store handles to subjects lists
-        self.subject_lists = {}
-        self.panels = {}
-
-        # Set layout for whole widget
-        self.layout = QtWidgets.QGridLayout()
-        self.layout.setContentsMargins(0,0,0,0)
-        self.layout.setSpacing(0)
-        self.setLayout(self.layout)
-
-        self.init_ui()
-
-        # self.setSizePolicy(QtWidgets.QSizePolicy.Maximum,QtWidgets.QSizePolicy.Maximum)
-        self.setSizePolicy(QtWidgets.QSizePolicy.Expanding,QtWidgets.QSizePolicy.Expanding)
-
-        self.setStyleSheet(styles.CONTROL_PANEL)
-
-    def init_ui(self):
-        """
-        Called on init, creates the UI components.
-
-        Specifically, for each pilot in :py:attr:`.pilots`,
-        make a :class:`subject_List`: and :class:`Pilot_Panel`:,
-        set size policies and connect Qt signals.
-        """
-        self.layout.setColumnStretch(0, 2)
-        self.layout.setColumnStretch(1, 2)
-
-        for pilot_id, pilot_params in self.pilots.items():
-            self.add_pilot(pilot_id, pilot_params.get('subjects', []))
-
-    @gui_event
-    def add_pilot(self, pilot_id:str, subjects:typing.Optional[list]=None):
-        """
-        Add a :class:`.Pilot_Panel` for a new pilot, and populate a :class:`.Subject_List` for it
-        Args:
-         pilot_id (str): ID of new pilot
-         subjects (list): Optional, list of any subjects that the pilot has.
-        Returns:
-        """
-        if subjects is None:
-            subjects = []
-
-        # Make a list of subjects
-        subject_list = Subject_List(subjects, drop_fn=self.update_db)
-        subject_list.setSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Expanding)
-        # subject_list.itemDoubleClicked.connect(self.edit_params)
-        self.subject_lists[pilot_id] = subject_list
-
-        # Make a panel for pilot control
-        pilot_panel = Pilot_Panel(pilot_id, subject_list, self.start_fn, self.ping_fn, self.create_subject)
-        pilot_panel.setFixedWidth(150)
-        self.panels[pilot_id] = pilot_panel
-
-        row_idx = self.layout.rowCount()
-
-        self.layout.addWidget(pilot_panel, row_idx, 1, 1, 1)
-        self.layout.addWidget(subject_list, row_idx, 2, 1, 1)
-
-    def create_subject(self, pilot):
-        """
-        Becomes :py:attr:`.Pilot_Panel.create_fn`.
-        Opens a :py:class:`.New_Subject_Wizard` to create a new subject file and assign protocol.
-        Finally, adds the new subject to the :py:attr:`~.Control_Panel.pilots` database and updates it.
-
-        Args:
-            pilot (str): Pilot name passed from :py:class:`.Pilot_Panel`, added to the created Subject object.
-        """
-        new_subject_wizard = New_Subject_Wizard()
-        new_subject_wizard.exec_()
-
-        # If the wizard completed successfully, get its values
-        if new_subject_wizard.result() == 1:
-
-            biography_vals = new_subject_wizard.bio_tab.values
-            self.logger.debug(f'subject wizard exited with 1, got biography vals {biography_vals}')
-            # TODO: Make a "session" history table that stashes pilot, git hash, step, etc. for each session - subjects might run on different pilots
-            biography_vals['pilot'] = pilot
-
-            # Make a new subject object, make it temporary because we want to close it
-            subject_obj = Subject(biography_vals['id'], new=True,
-                                biography=biography_vals)
-            self.subjects[biography_vals['id']] = subject_obj
-
-            # If a protocol was selected in the subject wizard, assign it.
-            try:
-                protocol_vals = new_subject_wizard.task_tab.values
-                if 'protocol' in protocol_vals.keys() and 'step' in protocol_vals.keys():
-                    protocol_file = os.path.join(prefs.get('PROTOCOLDIR'), protocol_vals['protocol'] + '.json')
-                    subject_obj.assign_protocol(protocol_file, int(protocol_vals['step']))
-                    self.logger.debug(f'assigned protocol with {protocol_vals}')
-                else:
-                    self.logger.warning(f'protocol couldnt be assigned, no step and protocol keys in protocol_vals.\ngot protocol_vals: {protocol_vals}')
-            except Exception as e:
-                self.logger.exception(f'exception when assigning protocol, continuing subject creation. \n{e}')
-                # the wizard couldn't find the protocol dir, so no task tab was made
-                # or no task was assigned
-
-            # Add subject to pilots dict, update it and our tabs
-            self.pilots[pilot]['subjects'].append(biography_vals['id'])
-            self.subject_lists[pilot].addItem(biography_vals['id'])
-            self.update_db()
-
-    # TODO: fix this
-    # def edit_params(self, item):
-    #     """
-    #     Args:
-    #         item:
-    #     """
-    #     # edit a subject's task parameters, called when subject double-clicked
-    #     subject = item.text()
-    #     if subject not in self.subjects.keys():
-    #         self.subjects[subject] = Subject(subject)
-    #
-    #     if '/current' not in self.subjects[subject].h5f:
-    #         Warning("Subject {} has no protocol!".format(subject))
-    #         return
-    #
-    #     protocol = self.subjects[subject].current
-    #     step = self.subjects[subject].step
-    #
-    #     protocol_edit = Protocol_Parameters_Dialogue(protocol, step)
-    #     protocol_edit.exec_()
-    #
-    #     if protocol_edit.result() == 1:
-    #         param_changes = protocol_edit.step_changes
-    #         # iterate through steps, checking for changes
-    #         for i, step_changes in enumerate(param_changes):
-    #             # if there are any changes to this step, stash them
-    #             if step_changes:
-    #                 for k, v in step_changes.items():
-    #                     self.subjects[subject].update_history('param', k, v, step=i)
-    #
-    #
-
-    def update_db(self, pilots:typing.Optional[dict]=None, **kwargs):
-        """
-        Gathers any changes in :class:`Subject_List` s and dumps :py:attr:`.pilots` to :py:attr:`.prefs.get('PILOT_DB')`
-
-        Args:
-            kwargs: Create new pilots by passing a dictionary with the structure
-
-                `new={'pilot_name':'pilot_values'}`
-
-                where `'pilot_values'` can be nothing, a list of subjects,
-                or any other information included in the pilot db
-        """
-        # if we were given a new pilot, add it
-        if 'new' in kwargs.keys():
-            for pilot, value in kwargs['new'].items():
-                self.pilots[pilot] = value
-
-        if pilots is None:
-            pilots = self.pilots.copy()
-
-        # gather subjects from lists
-        for pilot, mlist in self.subject_lists.items():
-            subjects = []
-            for i in range(mlist.count()):
-                subjects.append(mlist.item(i).text())
-
-            pilots[pilot]['subjects'] = subjects
-
-        # strip any state that's been stored
-        for p, val in pilots.items():
-            if 'state' in val.keys():
-                del val['state']
-
-        with open(prefs.get('PILOT_DB'), 'w') as pilot_file:
-            json.dump(self.pilots, pilot_file, indent=4, separators=(',', ': '))
-
 ####################################
 # Control Panel Widgets
 ###################################
-
-class Subject_List(QtWidgets.QListWidget):
-    """
-    A trivial modification of :class:`~.QtWidgets.QListWidget` that updates
-    :py:attr:`~.Terminal.pilots` when an item in the list is dragged to another location.
-
-    Should not be initialized except by :class:`.Control_Panel` .
-
-    Attributes:
-        subjects (list): A list of subjects ID's passed by :class:`.Control_Panel`
-        drop_fn (:py:meth:`.Control_Panel.update_db`): called on a drop event
-    """
-
-    def __init__(self, subjects=None, drop_fn=None):
-        """
-        Args:
-            subjects: see :py:attr:`~.Subject_List.subjects`. Can be `None` for an empty list
-            drop_fn: see :py:meth:`~.Subject_List.drop_fn`. Passed from :class:`.Control_Panel`
-        """
-        super(Subject_List, self).__init__()
-
-        # if we are passed a list of subjects, populate
-        if subjects:
-            self.subjects = subjects
-            self.populate_list()
-        else:
-            self.subjects = []
-
-        # make draggable
-        self.setDragDropMode(QtWidgets.QAbstractItemView.InternalMove)
-        self.setDragDropOverwriteMode(False)
-        self.setAcceptDrops(True)
-
-        # drop_fn gets called on a dropEvent (after calling the superclass method)
-        self.drop_fn = drop_fn
-
-    def populate_list(self):
-        """
-        Adds each item in :py:attr:`Subject_List.subjects` to the list.
-        """
-        for m in self.subjects:
-            self.addItem(m)
-
-    def dropEvent(self, event):
-        """
-        A trivial redefinition of :py:meth:`.QtWidgets.QListWidget.dropEvent`
-        that calls the parent `dropEvent` and then calls :py:attr:`~.Subject_List.drop_fn`
-
-        Args:
-            event: A :class:`.QtCore.QEvent` simply forwarded to the superclass.
-        """
-        # call the parent dropEvent to make sure all the list ops happen
-        super(Subject_List, self).dropEvent(event)
-        # then we call the drop_fn passed to us
-        self.drop_fn()
-
-
-class Pilot_Panel(QtWidgets.QWidget):
-    """
-    A little panel with
-
-    * the name of a pilot,
-    * A :class:`Pilot_Button` to start and stop the task
-    * Add and remove buttons to :py:meth:`~Pilot_Panel.create_subject` and :py:meth:`Pilot_Panel.remove_subject`
-
-    Note:
-        This class should not be instantiated except by :class:`Control_Panel`
-
-    Args:
-        pilot (str): The name of the pilot this panel controls
-        subject_list (:py:class:`.Subject_List`): The :py:class:`.Subject_List` we control
-        start_fn (:py:meth:`~autopilot.core.terminal.Terminal.toggle_start`): Passed by :class:`Control_Panel`
-        create_fn (:py:meth:`Control_Panel.create_subject`): Passed by :class:`Control_Panel`
-
-    Attributes:
-        layout (:py:class:`QtWidgets.QGridLayout`): Layout for UI elements
-        button (:class:`.Pilot_Button`): button used to control a pilot
-    """
-    def __init__(self, pilot=None, subject_list=None, start_fn=None, ping_fn=None, create_fn=None):
-        """
-
-        """
-        super(Pilot_Panel, self).__init__()
-
-        self.layout = QtWidgets.QGridLayout()
-        self.layout.setContentsMargins(0,0,0,0)
-        self.layout.setSpacing(0)
-        self.setLayout(self.layout)
-
-        self.pilot = pilot
-        self.subject_list = subject_list
-        self.start_fn = start_fn
-        self.ping_fn = ping_fn
-        self.create_fn = create_fn
-        self.button = None
-
-        self.init_ui()
-
-    def init_ui(self):
-        """
-        Initializes UI elements - creates widgets and adds to :py:attr:`Pilot_Panel.layout` .
-        Called on init.
-        """
-
-        label = QtWidgets.QLabel(self.pilot)
-        label.setStyleSheet("font: bold 14pt; text-align:right")
-        label.setAlignment(QtCore.Qt.AlignVCenter)
-        self.button = Pilot_Button(self.pilot, self.subject_list, self.start_fn, self.ping_fn)
-        add_button = QtWidgets.QPushButton("+")
-        add_button.clicked.connect(self.create_subject)
-        add_button.setSizePolicy(QtWidgets.QSizePolicy.Expanding,QtWidgets.QSizePolicy.Expanding)
-        remove_button = QtWidgets.QPushButton("-")
-        remove_button.clicked.connect(self.remove_subject)
-        remove_button.setSizePolicy(QtWidgets.QSizePolicy.Expanding,QtWidgets.QSizePolicy.Expanding)
-
-        self.layout.addWidget(label, 0, 0, 1, 2)
-        self.layout.addWidget(self.button, 1, 0, 1, 2)
-        self.layout.addWidget(add_button, 2,0,1,1)
-        self.layout.addWidget(remove_button, 2,1,1,1)
-
-        self.layout.setRowStretch(0, 3)
-        self.layout.setRowStretch(1, 2)
-        self.layout.setRowStretch(2, 1)
-
-    def remove_subject(self):
-        """
-        Remove the currently selected subject in :py:attr:`Pilot_Panel.subject_list`,
-        and calls the :py:meth:`Control_Panel.update_db` method.
-        """
-
-        current_subject = self.subject_list.currentItem().text()
-        msgbox = QtWidgets.QMessageBox()
-        msgbox.setText("\n(only removes from pilot_db.json, data will not be deleted)".format(current_subject))
-
-        msgBox = QtWidgets.QMessageBox()
-        msgBox.setText("Are you sure you would like to remove {}?".format(current_subject))
-        msgBox.setInformativeText("'Yes' only removes from pilot_db.json, data will not be deleted")
-        msgBox.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
-        msgBox.setDefaultButton(QtWidgets.QMessageBox.No)
-        ret = msgBox.exec_()
-
-        if ret == QtWidgets.QMessageBox.Yes:
-
-            self.subject_list.takeItem(self.subject_list.currentRow())
-            # the drop fn updates the db
-            self.subject_list.drop_fn()
-
-    def create_subject(self):
-        """
-        Just calls :py:meth:`Control_Panel.create_subject` with our `pilot` as the argument
-        """
-        self.create_fn(self.pilot)
-
-
-class Pilot_Button(QtWidgets.QPushButton):
-    def __init__(self, pilot=None, subject_list=None, start_fn=None, ping_fn=None):
-        """
-        A subclass of (toggled) :class:`QtWidgets.QPushButton` that incorporates the style logic of a
-        start/stop button - ie. color, text.
-
-        Starts grayed out, turns green if contact with a pilot is made.
-
-        Args:
-            pilot (str): The ID of the pilot that this button controls
-            subject_list (:py:class:`.Subject_List`): The Subject list used to determine which
-                subject is starting/stopping
-            start_fn (:py:meth:`~autopilot.core.terminal.Terminal.toggle_start`): The final
-                resting place of the toggle_start method
-
-        Attributes:
-            state (str): The state of our pilot, reflected in our graphical properties.
-                Mirrors :attr:`~.pilot.Pilot.state` , with an additional "DISCONNECTED"
-                state for before contact is made with the pilot.
-        """
-        super(Pilot_Button, self).__init__()
-
-        ## GUI Settings
-        self.setCheckable(False)
-        self.setChecked(False)
-        self.setEnabled(True)
-
-        self.normal_stylesheet = (
-            "QPushButton {color:white; background-color: green}"
-            "QPushButton:checked {color:white; background-color: red}"
-            "QPushButton:disabled {color:black; background-color: gray}"
-        )
-
-        self.limbo_stylesheet = (
-            "QPushButton {color:black; background-color: gray}"
-        )
-
-        self.setStyleSheet(self.limbo_stylesheet)
-        # at start, set our text to no pilot and wait for the signal
-        self.setText("?PING?")
-
-        # keep track of our visual and functional state.
-        self.state = "DISCONNECTED"
-
-        # Normally buttons only expand horizontally, but these big ole ones....
-        self.setSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Expanding)
-
-        # What's yr name anyway?
-        self.pilot = pilot
-
-        # What subjects do we know about?
-        self.subject_list = subject_list
-
-
-
-        # Passed a function to toggle start from the control panel
-        self.start_fn = start_fn
-        self.ping_fn = ping_fn
-        # toggle_start has a little sugar on it before sending to control panel
-        # use the clicked rather than toggled signal, clicked only triggers on user
-        # interaction, toggle is whenever the state is toggled - so programmatically
-        # toggling the button - ie. responding to pilot state changes - double triggers.
-        self.clicked.connect(self.toggle_start)
-
-        self.logger = init_logger(self)
-
-
-    def toggle_start(self):
-        """
-        Minor window dressing to call the :py:meth:`~.Pilot_Button.start_fn` with the
-        appropriate pilot, subject, and whether the task is starting or stopping
-
-        """
-        # If we're stopped, start, and vice versa...
-
-        if self.state == "DISCONNECTED":
-            # ping our lil bebs
-            self.ping_fn(self.pilot)
-            return
-
-        try:
-            current_subject = self.subject_list.currentItem().text()
-        except AttributeError:
-            self.logger.warning('Start button clicked, but no subject selected')
-            return
-
-        toggled = self.isChecked()
-        if toggled is True: # ie button is already down, already running.
-
-            self.setText("STOP")
-            self.start_fn(True, self.pilot, current_subject)
-
-        else:
-            self.setText("START")
-            self.start_fn(False, self.pilot, current_subject)
-
-    @gui_event
-    def set_state(self, state):
-        """
-        Set the button's appearance and state
-
-        Args:
-            state (str): one of ``('IDLE', 'RUNNING', 'STOPPING', 'DISCONNECTED')
-
-        .. todo::
-
-            There is some logic duplication in this class, ie. if the button state is changed
-            it also emits a start/stop signal to the pi, which is undesirable. This class needs
-            to be reworked.
-
-        Returns:
-
-        """
-        # if we're good, do nothing.
-        if state == self.state:
-            return
-
-
-        if state == "IDLE":
-            # responsive and waiting
-            self.setCheckable(True)
-            self.setEnabled(True)
-            self.setText('START')
-            self.setChecked(False)
-        elif state == "RUNNING":
-            # running a task
-            self.setCheckable(True)
-            self.setEnabled(True)
-            self.setText('STOP')
-            self.setChecked(True)
-        elif state == "STOPPING":
-            # stopping
-            self.setCheckable(True)
-            self.setEnabled(False)
-            self.setText("STOPPING")
-            self.setChecked(False)
-        elif state == "DISCONNECTED":
-            # contact w the pi is missing or lost
-            self.setCheckable(False)
-            self.setEnabled(True)
-            self.setText("?PING?")
-            self.setChecked(False)
-
-        if state == "DISCONNECTED":
-            self.setStyleSheet(self.limbo_stylesheet)
-        else:
-            if self.isChecked():
-                self.setText('STOP')
-            else:
-                self.setText('START')
-            self.setStyleSheet(self.normal_stylesheet)
-
-        self.state = state
 
 
 ##################################
@@ -631,1028 +62,6 @@ class Pilot_Button(QtWidgets.QPushButton):
 
 # TODO: Change these classes to use the update params windows
 
-class New_Subject_Wizard(QtWidgets.QDialog):
-    """
-    A popup that prompts you to define variables for a new :class:`.subject.Subject` object
-
-    Called by :py:meth:`.Control_Panel.create_subject` , which handles actually creating
-    the subject file and updating the :py:attr:`.Terminal.pilots` dict and file.
-
-    Contains two tabs
-    - :class:`~.New_Subject_Wizard.Biography_Tab` - to set basic biographical information about a subject
-    - :class:`~.New_Subject_Wizard.Task_Tab` - to set the protocol and step to start the subject on
-
-    Attributes:
-        protocol_dir (str): A full path to where protocols are stored,
-            received from :py:const:`.prefs.get('PROTOCOLDIR')`
-        bio_tab (:class:`~.New_Subject_Wizard.Biography_Tab`): Sub-object to set and store biographical variables
-        task_tab (:class:`~.New_Subject_Wizard.Task_Tab`): Sub-object to set and store protocol and step assignment
-    """
-
-    def __init__(self):
-        QtWidgets.QDialog.__init__(self)
-
-        self.protocol_dir = prefs.get('PROTOCOLDIR')
-
-        tabWidget = QtWidgets.QTabWidget()
-
-        self.bio_tab = self.Biography_Tab()
-        tabWidget.addTab(self.bio_tab, "Biography")
-
-        if self.protocol_dir:
-            self.task_tab = self.Task_Tab()
-            tabWidget.addTab(self.task_tab, "Protocol")
-
-        buttonBox = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        buttonBox.accepted.connect(self.accept)
-        buttonBox.rejected.connect(self.reject)
-
-        mainLayout = QtWidgets.QVBoxLayout()
-        mainLayout.addWidget(tabWidget)
-        mainLayout.addWidget(buttonBox)
-        self.setLayout(mainLayout)
-
-        self.setWindowTitle("Setup New Subject")
-
-    class Biography_Tab(QtWidgets.QWidget):
-        """
-        A widget that allows defining basic biographical attributes about a subject
-
-        Creates a set of widgets connected to :py:meth:`~.Biography_Tab.update_return_dict` that stores the parameters.
-
-        Warning:
-            The below attributes are **not** the object attributes, but are descriptions of the parameters
-            available in the values dictionary. The attributes themselves are PySide Widgets that set the values.
-
-        Attributes:
-            id (str): A Subject's ID or name
-            start_date (str): The date the subject started the task. Automatically filled by
-                :py:meth:`datetime.date.today().isoformat()`
-            blmass (float): The subject's baseline mass
-            minmass_pct (int): The percentage of baseline mass that a water restricted subject is allowed to reach
-            minmass (float): The subject's minimum mass, automatically calculated `blmass * (minmass_pct / 100.)`
-            genotype (str): A string describing the subject's genotype
-            expt (str): A tag to describe what experiment this subject is a part of
-        """
-        def __init__(self):
-            QtWidgets.QWidget.__init__(self)
-
-            # Input Labels
-            ID_label = QtWidgets.QLabel("ID:")
-            start_label = QtWidgets.QLabel("Start Date:")
-            blmass_label = QtWidgets.QLabel("Baseline Mass:")
-            minmasspct_label = QtWidgets.QLabel("% of Baseline Mass:")
-            minmass_label = QtWidgets.QLabel("Minimum Mass:")
-            genotype_label = QtWidgets.QLabel("Genotype:")
-            expt_label = QtWidgets.QLabel("Experiment Tag:")
-
-            # Input widgets
-            self.id = QtWidgets.QLineEdit()
-            self.start_date = QtWidgets.QLineEdit(datetime.date.today().isoformat())
-            self.blmass = QtWidgets.QLineEdit()
-            self.blmass.setValidator(QtGui.QDoubleValidator(0.0, 30.0, 1, self.blmass))
-            self.minmass_pct = QtWidgets.QSpinBox()
-            self.minmass_pct.setRange(0,100)
-            self.minmass_pct.setSingleStep(5)
-            self.minmass_pct.setSuffix('%')
-            self.minmass_pct.setValue(80)
-            self.minmass = QtWidgets.QLineEdit()
-            self.minmass.setValidator(QtGui.QDoubleValidator(0.0, 30.0, 1, self.minmass))
-            self.genotype = QtWidgets.QLineEdit()
-            self.expt     = QtWidgets.QLineEdit()
-
-            # Set return dictionary signals
-            self.id.editingFinished.connect(lambda: self.update_return_dict('id', self.id.text()))
-            self.start_date.editingFinished.connect(lambda: self.update_return_dict('start_date', self.start_date.text()))
-            self.blmass.editingFinished.connect(lambda: self.update_return_dict('baseline_mass', self.blmass.text()))
-            self.minmass.editingFinished.connect(lambda: self.update_return_dict('min_mass', self.minmass.text()))
-            self.genotype.editingFinished.connect(lambda: self.update_return_dict('genotype', self.genotype.text()))
-            self.expt.editingFinished.connect(lambda: self.update_return_dict('experiment', self.expt.text()))
-
-            # Set update minmass signals
-            self.blmass.editingFinished.connect(self.calc_minmass)
-            self.minmass_pct.valueChanged.connect(self.calc_minmass)
-
-            # Setup Layout
-            mainLayout = QtWidgets.QVBoxLayout()
-            mainLayout.addWidget(ID_label)
-            mainLayout.addWidget(self.id)
-            mainLayout.addWidget(start_label)
-            mainLayout.addWidget(self.start_date)
-            mainLayout.addWidget(blmass_label)
-            mainLayout.addWidget(self.blmass)
-            mainLayout.addWidget(minmasspct_label)
-            mainLayout.addWidget(self.minmass_pct)
-            mainLayout.addWidget(minmass_label)
-            mainLayout.addWidget(self.minmass)
-            mainLayout.addWidget(genotype_label)
-            mainLayout.addWidget(self.genotype)
-            mainLayout.addWidget(expt_label)
-            mainLayout.addWidget(self.expt)
-            mainLayout.addStretch(1)
-
-            self.setLayout(mainLayout)
-
-            # Dictionary to return values
-            self.values = {}
-
-        def update_return_dict(self, key, val):
-            """
-            Called by lambda functions by the widgets, eg.::
-
-                self.id.editingFinished.connect(lambda: self.update_return_dict('id', self.id.text()))
-
-            Args:
-                key (str): The key of the value being stored
-                val: The value being stored.
-            """
-            self.values[key] = val
-            # When values changed, update return dict
-
-        def calc_minmass(self):
-            """
-            Calculates the minimum mass for a subject based on its baseline mass
-            and the allowable percentage of that baseline
-            """
-            # minimum mass automatically from % and baseline
-            baseline = float(self.blmass.text())
-            pct = float(self.minmass_pct.text()[:-1])/100
-            self.minmass.setText(str(baseline*pct))
-
-    class Task_Tab(QtWidgets.QWidget):
-        """
-        A tab for selecting a task and step to assign to the subject.
-
-        Reads available tasks from `prefs.get('PROTOCOLDIR')` , lists them, and
-        creates a spinbox to select from the available steps.
-
-        Warning:
-            Like :class:`.Biography_Tab` , these are not the actual instance attributes.
-            Values are stored in a `values` dictionary.
-
-        Attributes:
-            protocol (str): the name of the assigned protocol, filename without .json extension
-            step (int): current step to assign.
-        """
-        def __init__(self):
-            QtWidgets.QWidget.__init__(self)
-
-            self.protocol_dir = prefs.get('PROTOCOLDIR')
-
-            topLabel = QtWidgets.QLabel("Protocols:")
-
-            # List available protocols
-            protocol_list = os.listdir(self.protocol_dir)
-            protocol_list = [os.path.splitext(p)[0] for p in protocol_list]
-
-            self.protocol_listbox = QtWidgets.QListWidget()
-            self.protocol_listbox.insertItems(0, protocol_list)
-            self.protocol_listbox.currentItemChanged.connect(self.protocol_changed)
-
-            # Make Step combobox
-            self.step_selection = QtWidgets.QComboBox()
-            self.step_selection.currentIndexChanged.connect(self.step_changed)
-
-            layout = QtWidgets.QVBoxLayout()
-            layout.addWidget(topLabel)
-            layout.addWidget(self.protocol_listbox)
-            layout.addWidget(self.step_selection)
-
-            self.setLayout(layout)
-
-            # Dict to return values
-            self.values = {}
-
-        def update_step_box(self):
-            """
-            Clears any steps that might be in the step selection box,
-            loads the protocol file and repopulates it.
-            """
-            # Clear box
-            while self.step_selection.count():
-                self.step_selection.removeItem(0)
-
-            # Load the protocol and parse its steps
-            protocol_str = self.protocol_listbox.currentItem().text()
-            protocol_file = os.path.join(self.protocol_dir,protocol_str + '.json')
-            with open(protocol_file) as protocol_file_open:
-                protocol = json.load(protocol_file_open)
-
-            step_list = []
-            self.step_ind   = {}
-            for i, s in enumerate(protocol):
-                step_list.append(s['step_name'])
-                self.step_ind[s['step_name']] = i
-
-            self.step_selection.insertItems(0, step_list)
-            self.step_selection.setCurrentIndex(0)
-
-        def protocol_changed(self):
-            """
-            When the protocol is changed, save the value and call :py:meth:`.update_step_box`.
-            """
-            self.values['protocol'] = self.protocol_listbox.currentItem().text()
-            self.update_step_box()
-
-        def step_changed(self):
-            """
-            When the step is changed, save it.
-            """
-            current_step = self.step_selection.currentText()
-            # Check that we have selected a step...
-            if current_step != '':
-                self.values['step'] = self.step_ind[current_step]
-
-
-class Protocol_Wizard(QtWidgets.QDialog):
-    """
-    A dialog window to create a new protocol.
-
-    Warning:
-        This is a heavily overloaded class, and will be split into separate objects
-        to handle parameters separately. For now this is what we got though and it works.
-
-    Protocols are collections of multiple tasks (steps)
-    with some graduation criterion for moving between them.
-
-    This widget is composed of three windows:
-
-    * **left**: possible task types from :func:`autopilot.get_task()`
-    * **center**: current steps in task
-    * **right**: :class:`.Parameters` for currently selected step.
-
-    The parameters that are used are of the form used by :py:attr:`.Task.PARAMS`
-    (see :py:attr:`.Nafc.PARAMS` for an example).
-
-    TODO:
-        Make specific parameter class so this definition is less squishy
-
-    its general structure is::
-
-        {'parameter_key': {'tag':'Human Readable Name',
-                           'type':'param_type'}}
-
-    while some parameter types have extra items, eg.::
-
-        {'list_param': {'tag':'Select from a List of Parameters',
-                        'type': 'list',
-                        'values': {'First Option':0, 'Second Option':1}}
-
-    where k:v pairs are still used with lists to allow parameter values (0, 1) be human readable.
-
-    The available types include:
-
-    * **int** - integer
-    * **float** - floating point number
-    * **bool** - boolean boolbox
-    * **list** - a list of `values` to choose from
-    * **sounds** - a :class:`.Sound_Widget` that allows sounds to be defined.
-    * **graduation** - a :class:`.Graduation_Widget` that allows graduation criteria to be defined
-
-    Attributes:
-        task_list (:class:`QtWidgets.QListWidget`): The leftmost window, lists available tasks
-        step_list (:class:`QtWidgets.QListWidget`): The center window, lists tasks currently in protocol
-        param_layout (:class:`QtWidgets.QFormLayout`): The right window, allows changing available
-            parameters for currently selected step.
-        steps (list): A list of dictionaries defining the protocol.
-
-    """
-    def __init__(self):
-        QtWidgets.QDialog.__init__(self)
-
-        # Left Task List/Add Step Box
-        addstep_label = QtWidgets.QLabel("Add Step")
-        addstep_label.setFixedHeight(40)
-        self.task_list = QtWidgets.QListWidget()
-        self.task_list.insertItems(0, autopilot.get_names('task'))
-        self.add_button = QtWidgets.QPushButton("+")
-        self.add_button.setFixedHeight(40)
-        self.add_button.clicked.connect(self.add_step)
-
-        addstep_layout = QtWidgets.QVBoxLayout()
-        addstep_layout.addWidget(addstep_label)
-        addstep_layout.addWidget(self.task_list)
-        addstep_layout.addWidget(self.add_button)
-
-        # Center Step List Box
-        steplist_label = QtWidgets.QLabel("Step List")
-        steplist_label.setFixedHeight(40)
-        self.step_list = QtWidgets.QListWidget()
-        self.step_list.setDragDropMode(QtWidgets.QAbstractItemView.InternalMove)
-        self.step_list.selectionMode = QtWidgets.QAbstractItemView.SingleSelection
-        self.step_list.itemSelectionChanged.connect(self.populate_params)
-        self.list_model = self.step_list.model()
-        self.list_model.rowsMoved.connect(self.reorder_steps)
-        self.remove_step_button = QtWidgets.QPushButton('-')
-        self.remove_step_button.setFixedHeight(40)
-        self.remove_step_button.clicked.connect(self.remove_step)
-
-        steplist_layout = QtWidgets.QVBoxLayout()
-        steplist_layout.addWidget(steplist_label)
-        steplist_layout.addWidget(self.step_list)
-        steplist_layout.addWidget(self.remove_step_button)
-
-        # Right Parameter Definition Window
-        param_label = QtWidgets.QLabel("Step Parameters")
-        param_label.setFixedHeight(40)
-        self.param_layout = QtWidgets.QFormLayout()
-        param_frame = QtWidgets.QFrame()
-        param_frame.setLayout(self.param_layout)
-
-        param_box_layout = QtWidgets.QVBoxLayout()
-        param_box_layout.addWidget(param_label)
-        param_box_layout.addWidget(param_frame)
-
-        # Main Layout
-        frame_layout = QtWidgets.QHBoxLayout()
-        frame_layout.addLayout(addstep_layout, stretch=1)
-        frame_layout.addLayout(steplist_layout, stretch=1)
-        frame_layout.addLayout(param_box_layout, stretch=3)
-
-        buttonBox = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        buttonBox.accepted.connect(self.accept)
-        buttonBox.rejected.connect(self.reject)
-
-        main_layout = QtWidgets.QVBoxLayout()
-        main_layout.addLayout(frame_layout)
-        main_layout.addWidget(buttonBox)
-
-        self.setLayout(main_layout)
-        self.setWindowTitle("Make New Protocol")
-
-        # List to store dicts of steps and params
-        self.steps = []
-
-    def add_step(self):
-        """
-        Loads `PARAMS` from task object, adds base parameters to :py:attr:`.steps` list
-        """
-        task_type = self.task_list.currentItem().text()
-        new_item = QtWidgets.QListWidgetItem()
-        new_item.setText(task_type)
-        task_params = copy.deepcopy(autopilot.get_task(task_type).PARAMS)
-
-        # Add params that are non-task specific
-        # Name of task type
-        task_params['task_type'] = {'type':'label','value':task_type}
-        # Prepend name of step shittily
-        task_params_temp = odict()
-        task_params_temp['step_name'] = {'type':'str', 'tag':'Step Name', 'value':task_type}
-        task_params_temp.update(task_params)
-        task_params.clear()
-        task_params.update(task_params_temp)
-        # add graduation field
-        task_params['graduation'] = {'type':'graduation', 'tag':'Graduation Criterion', 'value':{}}
-
-        self.steps.append(task_params)
-        self.step_list.addItem(new_item)
-        self.step_list.setCurrentItem(new_item)
-
-    def rename_step(self):
-        """
-        When the step name widget's text is changed,
-        fire this function to update :py:attr:`.step_list` which updates
-         :py:attr:`.steps`
-        """
-        sender = self.sender()
-        sender_text = sender.text()
-        current_step = self.step_list.item(self.step_list.currentRow())
-        current_step.setText(sender_text)
-
-    def remove_step(self):
-        """
-        Remove step from :py:attr:`.step_list` and :py:attr:`.steps`
-        """
-        step_index = self.step_list.currentRow()
-        del self.steps[step_index]
-        self.step_list.takeItem(step_index)
-
-    def populate_params(self):
-        """
-        Calls :py:meth:`.clear_params` and then creates widgets to edit parameter values.
-        Returns:
-
-        """
-        self.clear_params()
-
-        # Get current item index
-        step_index = self.step_list.currentRow()
-        step_dict = self.steps[step_index]
-
-        # Iterate through params to make input widgets
-        for k, v in step_dict.items():
-            # Make Input Widget depending on type
-            # Each Input type needs a different widget type,
-            # and each widget type has different methods to get/change values, so we have to do this ugly
-            if v['type'] == 'int' or v['type'] == 'str' or v['type'] == 'float':
-                rowtag = QtWidgets.QLabel(v['tag'])
-                input_widget = QtWidgets.QLineEdit()
-                input_widget.setObjectName(k)
-                if v['type'] == 'int':
-                    input_widget.setValidator(QtGui.QIntValidator())
-                elif v['type'] == 'float':
-                    input_widget.setValidator(QtGui.QDoubleValidator())
-                input_widget.editingFinished.connect(self.set_param)
-                if 'value' in v.keys():
-                    input_widget.setText(v['value'])
-                elif v['type'] == 'str':
-                    self.steps[step_index][k]['value'] = ''
-                self.param_layout.addRow(rowtag,input_widget)
-
-            elif v['type'] == 'bool':
-                rowtag = QtWidgets.QLabel(v['tag'])
-                input_widget = QtWidgets.QCheckBox()
-                input_widget.setObjectName(k)
-                input_widget.stateChanged.connect(self.set_param)
-                if 'value' in v.keys():
-                    input_widget.setChecked(v['value'])
-                else:
-                    self.steps[step_index][k]['value'] = False
-                self.param_layout.addRow(rowtag, input_widget)
-
-            elif v['type'] == 'list':
-                rowtag = QtWidgets.QLabel(v['tag'])
-                input_widget = QtWidgets.QListWidget()
-                input_widget.setObjectName(k)
-                sorted_values = sorted(v['values'], key=v['values'].get)
-                input_widget.insertItems(0, sorted_values)
-                input_widget.itemSelectionChanged.connect(self.set_param)
-                if 'value' in v.keys():
-                    select_item = input_widget.item(v['value'])
-                    input_widget.setCurrentItem(select_item)
-                else:
-                    self.steps[step_index][k]['value'] = sorted_values[0]
-                self.param_layout.addRow(rowtag, input_widget)
-                self.steps[step_index][k]['value'] = False
-            elif v['type'] == 'sounds':
-                self.sound_widget = Sound_Widget()
-                self.sound_widget.setObjectName(k)
-                self.sound_widget.pass_set_param_function(self.set_sounds)
-                self.param_layout.addRow(self.sound_widget)
-                self.steps[step_index][k]['sounds'] = {}
-                if 'value' in v.keys():
-                    self.sound_widget.populate_lists(v['value'])
-            elif v['type'] == 'graduation':
-                self.grad_widget = Graduation_Widget()
-                self.grad_widget.setObjectName(k)
-                self.grad_widget.set_graduation = self.set_graduation
-                self.param_layout.addRow(self.grad_widget)
-                if 'type' in v['value'].keys():
-                    combo_index = self.grad_widget.type_selection.findText(v['value']['type'])
-                    self.grad_widget.type_selection.setCurrentIndex(combo_index)
-                    self.grad_widget.populate_params(v['value']['value'])
-
-            elif v['type'] == 'label':
-                # This is a .json label not for display
-                pass
-
-            # Step name needs to be hooked up to the step list text
-            if k == 'step_name':
-                input_widget.editingFinished.connect(self.rename_step)
-
-        # TODO: Implement dependencies between parameters
-
-    def clear_params(self):
-        """
-        Clears widgets from parameter window
-        """
-        while self.param_layout.count():
-            child = self.param_layout.takeAt(0)
-            if child.widget():
-                child.widget().deleteLater()
-
-    def reorder_steps(self, *args):
-        """
-        When steps are dragged into a different order, update the step dictionary
-
-        Args:
-            *args: Input from our :py:attr:`.step_list` 's :class:`.QtWidgets.QListModel` 's reorder signal.
-        """
-        # arg positions 1 and 4 are starting and ending positions in the list, respectively
-        # We reorder our step list so the params line up.
-        before = args[1]
-        after = args[4]
-        self.steps.insert(after, self.steps.pop(before))
-
-    def set_param(self):
-        """
-        Callback function connected to the signal each widget uses to signal it has changed.
-
-        Identifies the param that was changed, gets the current value, and updates `self.steps`
-        """
-        sender = self.sender()
-        param_name = sender.objectName()
-        current_step = self.step_list.currentRow()
-        sender_type = self.steps[current_step][param_name]['type']
-
-        if sender_type == 'bool':
-            self.steps[current_step][param_name]['value'] = sender.isChecked()
-        elif sender_type == 'list':
-            list_text = sender.currentItem().text()
-            #list_value = self.steps[current_step][param_name]['values'][list_text]
-            self.steps[current_step][param_name]['value'] = list_text
-        elif sender_type == 'sounds':
-            self.steps[current_step][param_name]['value'] = self.sound_widget.sound_dict
-        else:
-            try:
-                sender_text = ast.literal_eval(sender.text())
-            except:
-                sender_text = sender.text()
-            self.steps[current_step][param_name]['value'] = sender_text
-
-    def set_sounds(self):
-        """
-        Stores parameters that define sounds.
-
-        Sound parameters work a bit differently, specifically we have to retrieve
-        :py:attr:`.Sound_Widget.sound_dict`.
-        """
-        current_step = self.step_list.currentRow()
-        #if 'sounds' in self.steps[current_step]['stim'].keys():
-        #    self.steps[current_step][param_name]['sounds']['value'].update(self.sound_widget.sound_dict)
-        #else:
-        self.steps[current_step]['stim']['sounds'] = self.sound_widget.sound_dict
-
-    def set_graduation(self):
-        """
-        Stores parameters that define graduation criteria in `self.steps`
-
-        Graduation parameters work a bit differently, specifically we have to retrieve
-        :py:attr:`.Graduation_Widget.param_dict`.
-        """
-        current_step = self.step_list.currentRow()
-        grad_type = self.grad_widget.type
-        grad_params = self.grad_widget.param_dict
-        self.steps[current_step]['graduation']['value'] = {'type':grad_type,'value':grad_params}
-
-
-    def check_depends(self):
-        """
-        Handle dependencies between parameters, eg. if "correction trials" are unchecked,
-        the box that defines the correction trial percentage should be grayed out.
-
-        TODO:
-            Not implemented.
-        """
-        # TODO: Make dependent fields unavailable if dependencies unmet
-        # I mean if it really matters
-        pass
-
-class Graduation_Widget(QtWidgets.QWidget):
-    """
-    A widget used in :class:`.Protocol_Wizard` to define graduation parameters.
-
-    See :py:mod:`.tasks.graduation` .
-
-    A protocol is composed of multiple tasks (steps), and graduation criteria
-    define when a subject should progress through those steps.
-
-    eg. a subject should graduate one stage after 300 trials, or after it reaches
-    75% accuracy over the last 500 trials.
-
-    Attributes:
-        type_selection (:class:`QtWidgets.QComboBox`): A box to select from the available
-            graduation types listed in :func:`autopilot.get_task()` . Has its `currentIndexChanged`
-            signal connected to :py:meth:`.Graduation_Widget.populate_params`
-        param_dict (dict): Stores the type of graduation and the relevant params,
-            fetched by :class:`.Protocol_Wizard` when defining a protocol.
-        set_graduation (:py:meth:`.Protocol_Wizard.set_graduation`): Passed to us after we're inited.
-    """
-
-    def __init__(self):
-        super(Graduation_Widget, self).__init__()
-
-        # Grad type dropdown
-        type_label = QtWidgets.QLabel("Graduation Criterion:")
-        self.type_selection = QtWidgets.QComboBox()
-        self.type_selection.insertItems(0, autopilot.get_names('graduation'))
-        self.type_selection.currentIndexChanged.connect(self.populate_params)
-
-        # Param form
-        self.param_layout = QtWidgets.QFormLayout()
-
-        layout = QtWidgets.QVBoxLayout()
-        layout.addWidget(type_label)
-        layout.addWidget(self.type_selection)
-        layout.addLayout(self.param_layout)
-
-        self.setLayout(layout)
-
-        self.param_dict = {}
-
-        self.type = ""
-
-        self.set_graduation = None
-
-        self.populate_params()
-
-    def populate_params(self, params=None):
-        """
-        Repopulate the widget with fields to edit graduation parameters, fill fields
-        if we are passed `params`.
-
-        Each :class:`QtWidgets.QLineEdit` 's :py:meth:`.QLineEdit.editingFinished` signal is connected
-        to :py:meth:`.Graduation_Widget.store_param` .
-
-        TODO:
-            For now we assume all parameters are defined with a text edit box, so it's not
-            clear how we'd do boolean parameters for example. This will be fixed with refactoring
-            the parameter scheme.
-
-        Args:
-            params (dict): In the case that :class:`.Protocol_Wizard` switches us back to
-                a step where we have already defined graduation parameters, it will pass them
-                so we can repopulate the relevant widgets with them.
-        """
-        self.clear_params()
-        self.type = self.type_selection.currentText()
-        self.param_dict['type'] = self.type
-
-        for k in autopilot.get_task(self.type).PARAMS:
-            edit_box = QtWidgets.QLineEdit()
-            edit_box.setObjectName(k)
-            edit_box.editingFinished.connect(self.store_param)
-            if isinstance(params, dict):
-                if k in params.keys():
-                    edit_box.setText(params[k])
-            self.param_layout.addRow(QtWidgets.QLabel(k), edit_box)
-
-    def clear_params(self):
-        """
-        Clear any parameter widgets we have.
-        """
-        while self.param_layout.count():
-            child = self.param_layout.takeAt(0)
-            if child.widget():
-                child.widget().deleteLater()
-
-    def store_param(self):
-        """
-        When a parameter is edited, save it in our param_dict, and also call our
-        `set_graduation` method, which should be :py:meth:`.Protocol_Wizard.set_graduation`
-        passed to us after instantiation.
-
-        If we were not passed `set_graduation`, just saves in `param_dict`.
-        """
-        sender = self.sender()
-        name = sender.objectName()
-        self.param_dict[name] = sender.text()
-
-        if not callable(self.set_graduation):
-            Warning("Stored Graduation parameters in our param_dict, but wasn't passed a set_graduation method!")
-            return
-
-        self.set_graduation()
-
-class Drag_List(QtWidgets.QListWidget):
-    """
-    A :class:`QtWidgets.QListWidget` that is capable of having files dragged & dropped.
-
-    copied with much gratitude from `stackoverflow <https://stackoverflow.com/a/25614674>`_
-
-    Primarily used in :class:`.Sound_Widget` to be able to drop sound files.
-
-    To use: connect `fileDropped` to a method, that method will receive a list of files
-    dragged onto this widget.
-
-    Attributes:
-        fileDropped (:class:`QtCore.Signal`): A Qt signal that takes a list
-    """
-    fileDropped = QtCore.Signal(list)
-
-    def __init__(self):
-        # type: () -> None
-        super(Drag_List, self).__init__()
-        self.setAcceptDrops(True)
-
-    def dragEnterEvent(self, e):
-        """
-        When files are dragged over us, if they have paths in them,
-        accept the event.
-
-        Args:
-            e (:class:`QtCore.QEvent`): containing the drag information.
-        """
-        if e.mimeData().hasUrls:
-            e.accept()
-        else:
-            e.ignore()
-
-    def dragMoveEvent(self, event):
-        """
-        If the `dragEnterEvent` was accepted, while the drag is being moved within us,
-        `setDropAction` to :class:`.QtCore.Qt.CopyAction`
-
-        Args:
-            event (:class:`QtCore.QEvent`): containing the drag information.
-        """
-        if event.mimeData().hasUrls:
-            event.setDropAction(QtCore.Qt.CopyAction)
-            event.accept()
-        else:
-            event.ignore()
-
-    def dropEvent(self, event):
-        """
-        When the files are finally dropped, if they contain paths,
-        emit the list of paths through the `fileDropped` signal.
-
-        Args:
-            event (:class:`QtCore.QEvent`): containing the drag information.
-        """
-        if event.mimeData().hasUrls:
-            event.setDropAction(QtCore.Qt.CopyAction)
-            event.accept()
-            links = []
-            for url in event.mimeData().urls():
-                links.append(str(url.toLocalFile()))
-            self.fileDropped.emit(links)
-        else:
-            event.ignore()
-
-
-class Sound_Widget(QtWidgets.QWidget):
-    """
-    A widget that allows sounds to be parameterized.
-
-    Used in :class:`.Protocol_Wizard` .
-
-    Has two :class:`.Drag_List` s for left and right sounds (for a 2afc task), given
-    Buttons beneath them allow adding and removing sounds.
-
-    Adding a sound will open a :class:`.Add_SoundDialog`
-
-    TODO:
-        Sounds will eventually be more elegantly managed by a ... sound manager..
-        For now sound managers are rudimentary and only support random presentation
-        with correction trials and bias correction.
-
-    Attributes:
-        sound_dict (dict): Dictionary with the structure::
-
-                {'L': [{'param_1':'param_1', ... }], 'R': [...]}
-
-            where multiple sounds can be present in either 'L' or 'R' list.
-
-
-
-    """
-    def __init__(self):
-        # type: () -> None
-        QtWidgets.QWidget.__init__(self)
-
-        self.sounddir = prefs.get('SOUNDDIR')
-
-        self.set_sounds = None
-
-        # Left sounds
-        left_label = QtWidgets.QLabel("Left Sounds")
-        left_label.setFixedHeight(30)
-        self.left_list = Drag_List()
-        self.left_list.fileDropped.connect(self.files_dropped)
-        self.left_list.setObjectName("L")
-        self.add_left_button = QtWidgets.QPushButton("+")
-        self.add_left_button.setFixedHeight(30)
-        self.add_left_button.clicked.connect(lambda: self.add_sound('L'))
-        self.remove_left_button = QtWidgets.QPushButton("-")
-        self.remove_left_button.setFixedHeight(30)
-        self.remove_left_button.clicked.connect(lambda: self.remove_sound('L'))
-
-        left_layout = QtWidgets.QVBoxLayout()
-        left_button_layout = QtWidgets.QHBoxLayout()
-        left_button_layout.addWidget(self.add_left_button)
-        left_button_layout.addWidget(self.remove_left_button)
-        left_layout.addWidget(left_label)
-        left_layout.addWidget(self.left_list)
-        left_layout.addLayout(left_button_layout)
-
-        # Right sounds
-        right_label = QtWidgets.QLabel("Right Sounds")
-        right_label.setFixedHeight(30)
-        self.right_list = Drag_List()
-        self.right_list.fileDropped.connect(self.files_dropped)
-        self.right_list.setObjectName("R")
-        self.add_right_button = QtWidgets.QPushButton("+")
-        self.add_right_button.setFixedHeight(30)
-        self.add_right_button.clicked.connect(lambda: self.add_sound('R'))
-        self.remove_right_button = QtWidgets.QPushButton("-")
-        self.remove_right_button.setFixedHeight(30)
-        self.remove_right_button.clicked.connect(lambda: self.remove_sound('R'))
-
-        right_layout = QtWidgets.QVBoxLayout()
-        right_button_layout = QtWidgets.QHBoxLayout()
-        right_button_layout.addWidget(self.add_right_button)
-        right_button_layout.addWidget(self.remove_right_button)
-        right_layout.addWidget(right_label)
-        right_layout.addWidget(self.right_list)
-        right_layout.addLayout(right_button_layout)
-
-        self.sound_dict = {'L': [], 'R': []}
-
-        main_layout = QtWidgets.QHBoxLayout()
-        main_layout.addLayout(left_layout)
-        main_layout.addLayout(right_layout)
-        self.setLayout(main_layout)
-
-        # TODO:Add drag and drop for files
-
-    def pass_set_param_function(self, set_param_fnxn):
-        """
-        Receives :py:meth:`.Protocol_Wizard.set_sounds`
-
-        Args:
-            set_param_fnxn (:py:meth:`.Protocol_Wizard.set_sounds`): Called when sounds are changed.
-        """
-        self.set_sounds = set_param_fnxn
-
-    def add_sound(self, side):
-        """
-        When the "+" button on either side is pressed, open an :class:`.Add_Sound_Dialog`.
-
-        Args:
-            side (str): The buttons are connected with a lambda function, this will be either 'L' or 'R'.
-                Used to add sounds to the `sound_dict`
-        """
-        new_sound = self.Add_Sound_Dialog()
-        new_sound.exec_()
-
-        if new_sound.result() == 1:
-            self.sound_dict[side].append(new_sound.param_dict)
-            if side == 'L':
-                self.left_list.addItem(new_sound.param_dict['type'])
-            elif side == 'R':
-                self.right_list.addItem(new_sound.param_dict['type'])
-            self.set_sounds()
-
-    def remove_sound(self, side):
-        """
-        When the "-" button is pressed, remove the currently highlighted sound.
-
-        Args:
-            side (str): The buttons are connected with a lambda function, this will be either 'L' or 'R'.
-                Selects that list so we can remove the currently selected row.
-        """
-        if side == 'L':
-            current_sound = self.left_list.currentRow()
-            del self.sound_dict['L'][current_sound]
-            self.left_list.takeItem(current_sound)
-        elif side == 'R':
-            current_sound = self.right_list.currentRow()
-            del self.sound_dict['R'][current_sound]
-            self.right_list.takeItem(current_sound)
-        self.set_sounds()
-
-    def populate_lists(self, sound_dict):
-        """
-        Populates the sound lists after re-selecting a step.
-
-        Args:
-            sound_dict (dict): passed to us by :class:`.Protocol_Wizard` upon reselecting a step.
-        """
-        # Populate the sound lists after re-selecting a step
-        self.sound_dict = sound_dict
-        for k in self.sound_dict['L']:
-            self.left_list.addItem(k['type'])
-        for k in self.sound_dict['R']:
-            self.right_list.addItem(k['type'])
-
-    def files_dropped(self, files):
-        """
-        Warning:
-            This was programmed hastily and is pretty idiosyncratic to my use.
-
-            It does work for general files but has some extra logic built in to handle my stimuli.
-
-            To be made more general in v0.3
-
-        Note:
-            Sounds must be in the folder specified in `prefs.get('SOUNDDIR')`.
-
-        When files are dropped on the lists, strips `prefs.get('SOUNDDIR')` from them to make them
-        relative paths, adds them to the `sound_dict`
-
-        Args:
-            files (list): List of absolute paths.
-        """
-        # TODO: Make this more general...
-        msg = QtWidgets.QMessageBox()
-        msg.setText("Are these Speech sounds in the format '/speaker/cv/cv_#.wav'?")
-        msg.setStandardButtons(QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No)
-        ret = msg.exec_()
-
-        sender = self.sender()
-        side = sender.objectName()
-
-        if ret == QtWidgets.QMessageBox.No:
-            for f in files:
-                f = f.strip(self.sounddir)
-
-                self.sound_dict[side].append({'type':'File', 'path':f})
-                if side == 'L':
-                    self.left_list.addItem(f)
-                elif side == 'R':
-                    self.right_list.addItem(f)
-
-
-        elif ret == QtWidgets.QMessageBox.Yes:
-            for f in files:
-                f = f.strip(self.sounddir)
-                f_split = f.split(os.sep)
-                speaker = f_split[0]
-                cv = f_split[-1].split('.')[0].split('_')[0]
-                consonant = cv[0]
-                vowel = cv[1:]
-                token = f_split[-1].split('.')[0].split('_')[1]
-                param_dict = {'type':'Speech','path':f,
-                 'speaker':speaker,'consonant':consonant,
-                 'vowel':vowel,'token':token}
-                self.sound_dict[side].append(param_dict)
-                if side == 'L':
-                    self.left_list.addItem(f)
-                elif side == 'R':
-                    self.right_list.addItem(f)
-
-        self.set_sounds()
-
-    class Add_Sound_Dialog(QtWidgets.QDialog):
-        """
-        Presents a dialog to define a new sound.
-
-        Makes a selection box to choose the sound type from
-        ``autopilot.get_names('sound')``, and then populates edit boxes
-        so we can fill in its `PARAMS` .
-
-        Attributes:
-            type_selection (:class:`QtWidgets.QComboBox`): Select from a list of available sounds
-            param_dict (dict): Parameters that are retreived by the calling :class:`.Sound_Widget`.
-
-        """
-        def __init__(self):
-            # type: () -> None
-            QtWidgets.QDialog.__init__(self)
-
-            # Sound type dropdown
-            type_label = QtWidgets.QLabel("Sound Type:")
-            self.type_selection = QtWidgets.QComboBox()
-            self.type_selection.insertItems(0, autopilot.get_names('sound'))
-            self.type_selection.currentIndexChanged.connect(self.populate_params)
-
-            # Param form
-            self.param_layout = QtWidgets.QFormLayout()
-
-            # Button box
-            buttonBox = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-            buttonBox.accepted.connect(self.accept)
-            buttonBox.rejected.connect(self.reject)
-
-            # Layout
-            layout = QtWidgets.QVBoxLayout()
-            layout.addWidget(type_label)
-            layout.addWidget(self.type_selection)
-            layout.addLayout(self.param_layout)
-            layout.addWidget(buttonBox)
-
-            self.setLayout(layout)
-
-
-            # dict for storing params
-            self.param_dict = {}
-
-            # do initial population
-            self.populate_params()
-
-        def populate_params(self):
-            """
-            When a sound type is selected, make a :class:`.QtWidgets.QLineEdit` for each
-            `PARAM` in its definition.
-            """
-            self.clear_params()
-
-            self.type = self.type_selection.currentText()
-            self.param_dict['type'] = self.type
-
-            for k in autopilot.get('sound', self.type).PARAMS:
-                edit_box = QtWidgets.QLineEdit()
-                edit_box.setObjectName(k)
-                edit_box.editingFinished.connect(self.store_param)
-                self.param_layout.addRow(QtWidgets.QLabel(k), edit_box)
-
-        def clear_params(self):
-            """
-            Clear all current widgets
-            """
-            while self.param_layout.count():
-                child = self.param_layout.takeAt(0)
-                if child.widget():
-                    child.widget().deleteLater()
-
-        def store_param(self):
-            """
-            When one of our edit boxes is edited, stash the parameter in `param_dict`
-            """
-            sender = self.sender()
-            name = sender.objectName()
-            self.param_dict[name] = sender.text()
 
 ###################################3
 # Tools
