+++ autopilot/core/terminal.py
@@ -45,7 +45,9 @@ from autopilot.core.subject import Subject
 from autopilot.core.plots import Plot_Widget
 from autopilot.networking import Net_Node, Terminal_Station
 from autopilot.utils.invoker import get_invoker
-from autopilot.core.gui import Control_Panel, Protocol_Wizard, Weights, Reassign, Calibrate_Water, Bandwidth_Test, pop_dialog, Stream_Video, Plugins
+from autopilot.core.gui import Weights, Reassign, Calibrate_Water, Bandwidth_Test, pop_dialog, Stream_Video, Plugins
+from autopilot.gui.widgets.protocol import Protocol_Wizard
+from autopilot.gui.widgets.terminal import Control_Panel
 from autopilot.core.loggers import init_logger
 
 # Try to import viz, but continue if that doesn't work
