+++ autopilot/gui/__init__.py
@@ -1,0 +1,14 @@
+_MAPS = {
+    'dialog': {
+        'icon': {
+            'info': QtWidgets.QMessageBox.Information,
+            'question': QtWidgets.QMessageBox.Question,
+            'warning': QtWidgets.QMessageBox.Warning,
+            'error': QtWidgets.QMessageBox.Critical
+        },
+        'modality': {
+            'modal': QtCore.Qt.NonModal,
+            'nonmodal': QtCore.Qt.WindowModal
+        }
+    }
+}
